requirements.txt
Django~=5.0
psycopg2-binary~=2.9
gunicorn~=21.2
python-dotenv~=1.0
djangorestframework~=3.14
django-cors-headers~=4.3
djangorestframework-simplejwt~=5.3
openpyxl~=3.1
xlrd~=2.0
beautifulsoup4~=4.12
requests~=2.31

# Security hardening
django-csp~=3.8
django-ratelimit~=4.1

# Telegram Bot
aiogram~=3.4
aiohttp~=3.9

# Celery & Redis
celery~=5.3
redis~=5.0
celery[redis]

# Image processing
Pillow~=10.2
